Fixed edge case where other provider adds user before mp_auth does
If another provider is present on the system (such as the resource
server acting as a portal in addition to being a resource server),
it's possible for the user to already exist. This adds a check to
make sure we reuse the user if the username matches (usernames
must be unique in django).